refactor(collections): make zod the single source of truth for the schema contract
Move every zod definition out of collection/server/discovery.ts into a new
isomorphic core/schemaZ.ts, and derive ALL schema-contract TypeScript types
from it via type-only `z.infer` imports — deleting the three hand-maintained
mirrors (core/schema.ts interfaces, core/where.ts predicate shapes, feeds
ingestTypes.ts DeclarativeIngestSpec/AgentIngestSpec).

Why: the zod schemas and the interfaces were kept in sync by hand with no
`z.infer`/`satisfies` tie. Because z.object strips unknown keys, adding a key
to the interface but forgetting the zod mirror silently deleted it from every
parsed schema at runtime. Each planned collection-ontology feature
(plans/collection-ontology.md) adds new schema keys, so the foundation is
fixed first (plan step 0, Phase A).

Structure:
- FieldSpecZ / SubFieldSpecZ become z.discriminatedUnion("type", ...) — each
  variant declares only its own keys (ref requires `to`, enum requires
  `values`, toggle requires field/onValue/offValue, ...), replacing the flat
  optional-key bag + 7 chained refines. New field types land as new union
  members. All schema-level cross-field refines and messages are unchanged.
- core/ids.ts: pure slug / record-id patterns shared by the validator and
  server/paths.ts's CodeQL-sanitizer wrappers (same regexes, one home).
- templatePath.ts moves to core/ (it was already dependency-free by design);
  a server shim keeps existing import paths (skill-bridge) working.
- Browser bundle stays zod-free: schema.ts imports schemaZ types type-only
  (erased at emit); at runtime the dependency points the other way
  (schemaZ -> schema consts). Verified: dist/collection/index.js has no zod.
- discovery.ts shrinks to discovery logic only and re-exports
  CollectionSchemaZ so manageCollection putSchema / registry importWriter
  resolve it unchanged.

Fallout: 10 call sites needed explicit `field.type === "..."` narrows
(embedTargetId, draft.ts, enumColors.ts, two plugin composables, three Vue
components) plus test fixtures. Everything else already narrowed. One piece
of dead code surfaced and removed: derived-sort compared display ===
"datetime", a value the validator never admitted for derived fields.

Behavior notes: acceptance rules are byte-compatible; the observable
differences are (1) junk keys on the wrong variant are now stripped from
parsed output (putSchema writes parsed.data, so it drops such junk), and
(2) missing-variant-key validation messages take zod's typed shape instead
of the custom refine text.

Verified: full-repo typecheck (caches cleared), 1297 host + 385 core tests,
lint, format, monorepo build all green.
