+++ docs/shared-utils.md
@@ -91,6 +91,8 @@ This catalog only covers **cross-cutting** helpers — formatters, error helpers
 | `packages/core/src/collection/core/dynamicIcon.ts` (`@mulmoclaude/core/collection`) | `selectDynamicRecord(records, source, orderBy, recordsById?)` / `resolveIcon(record, spec, staticIcon, recordsById?)` / `firstDateField(schema)` | Pure resolver for `CollectionSchema.dynamicIcon` (data-driven launcher icon): pick the deciding source record, then map it through the first-match-wins `rules`. `recordsById` (source records keyed by primaryKey, default `{}`) resolves any `valueFrom` value reference inside a `where`. Server wraps it in `packages/core/src/collection/server/dynamicIcon.ts`'s `computeCollectionIcon(collection, opts?)` (fails soft to `fallback ?? schema.icon`). |
 | `packages/core/src/collection/core/where.ts` (`@mulmoclaude/core/collection`) | `matchesWhere(where, record, recordsById?)` | SQL-like AND-of-conditions predicate (`eq`/`ne`/`in`/`gt`/`gte`/`lt`/`lte`/`contains`) used by `dynamicIcon`'s `source.where` / `rules[].where`. A condition's value is a literal `value` or a `valueFrom: { record, field }` reference resolved against `recordsById` (default `{}`); an unresolved reference makes the condition false for every op. Distinct from `whenMatches`/`CollectionWhen` (single-field, used by fields/actions) — don't conflate the two. |
 | `packages/core/src/collection-watchers/clock.ts` (`@mulmoclaude/core/collection-watchers`) | `evalNow()` | Evaluation-only injectable clock: returns `new Date()` unless `MULMOCLAUDE_FAKE_NOW` is set to a parseable date, in which case that date (re-read per call). Used as the reconciler/tick `now` default so evaluation runs and tests advance trigger/spawn time deterministically. Never set the env var in normal operation. |
+| `packages/core/src/collection/core/schemaZ.ts` | `CollectionSchemaZ` / `FieldSpecZ` / `SubFieldSpecZ` / `WhenZ` / `IngestZ` / … | The zod SINGLE SOURCE OF TRUTH for the `schema.json` contract; every type in `core/schema.ts` (and `where.ts` / feeds `ingestTypes.ts`) is `z.infer`-derived from here. Isomorphic but deliberately not in the browser barrel — servers import `CollectionSchemaZ` via `collection/server` (discovery re-exports it); browser code imports only the derived types. Add a field type or action kind by adding a union VARIANT here, never a parallel interface. |
+| `packages/core/src/collection/core/ids.ts` (`@mulmoclaude/core/collection`) | `isSafeSlug(value)` / `isSafeRecordId(value)` / `SAFE_SLUG_PATTERN` / `SAFE_RECORD_ID_PATTERN` | Pure (node-free) slug / record-id character rules for VALIDATION. The server-side `safeSlugName`/`safeRecordId` in `collection/server/paths.ts` wrap the same patterns with the `path.basename` round-trip CodeQL needs — use those whenever the value builds a filesystem path. |
 
 ## Plugin Infrastructure
