+++ packages/core/src/collection/core/draft.ts
@@ -125,7 +125,7 @@ function isMissingDraftValue(value: unknown): boolean {
 
 /** Label of the first required table sub-field empty in any row, else null. */
 function firstMissingTableSubField(field: FieldSpec, rows: TableRowDraft[] | undefined): string | null {
-  if (!field.of || !rows) return null;
+  if (field.type !== "table" || !rows) return null;
   for (let rowIdx = 0; rowIdx < rows.length; rowIdx++) {
     const row = rows[rowIdx];
     for (const [subKey, subField] of Object.entries(field.of)) {
