+++ test/plugins/collection/test_linkedCollectionCaches.ts
@@ -19,7 +19,9 @@ import type {
 import type { CollectionApiResult } from "../../../packages/plugins/collection-plugin/src/vue/uiContext";
 import { fetchLinkedCaches, linkedTargets } from "../../../packages/plugins/collection-plugin/src/vue/useLinkedCollectionCaches";
 
-const field = (type: FieldType, extra: Partial<FieldSpec> = {}): FieldSpec => ({ type, label: type, ...extra });
+// Fixture builder: assembles a spec dynamically (often deliberately partial),
+// so it casts rather than satisfying the discriminated union per-variant.
+const field = (type: FieldType, extra: Record<string, unknown> = {}): FieldSpec => ({ type, label: type, ...extra }) as FieldSpec;
 
 const makeSchema = (fields: Record<string, FieldSpec>, primaryKey = "id"): CollectionSchema => ({
   title: "Test",
